@@ -19,56 +20,69 @@
     "font-awesome",
     "es6",
     "ES2015",
-    "babeljs",
+    "es7",
+    "babel",
     "webpack",
-    "express"
+    "hot",
+    "reload",
+    "hmr",
+    "express",
+    "postcss",
+    "cssnext"
   ],
-  "description": "This is a boilerplate for starting a project with ReactJS... Webpack Dev Server is used to serve the application in dev mode. Express is used for the api",
+  "description": "Get started with React, Redux, Express...",
   "engines": {
-    "node": ">= 0.12.0"
+    "node": ">=5.0.0",
+    "npm": "^3.3.12"
   },
   "files": [
     "CHANGELOG.md",
     "README.md"
   ],
   "dependencies": {
-    "body-parser": "^1.14.0",
-    "classnames": "^2.2.1",
-    "express": "^4.13.0",
+    "classnames": "^2.2.3",
+    "express": "^4.13.4",
     "font-awesome": "^4.5.0",
-    "framework7": "^1.3.1",
-    "history": "^1.12.6",
-    "jquery": "^2.1.4",
-    "lodash": "^3.10.0",
-    "mongodb": "^2.1.1",
-    "mongoose": "^4.2.9",
-    "react": "^0.14.2",
-    "react-dom": "^0.14.2",
-    "react-redux": "^4.0.1",
-    "react-router": "^1.0.0-rc3",
-    "redux": "^3.0.5",
-    "redux-router": "^1.0.0-beta5",
-    "superagent": "^1.6.1",
+    "framework7": "^1.4.0",
+    "lodash": "^4.3.0",
+    "react": "^0.14.7",
+    "react-dom": "^0.14.7",
+    "react-redux": "^4.4.0",
+    "react-router": "^2.0.0",
+    "react-router-redux": "^3.0.0",
+    "redux": "^3.3.1",
+    "superagent": "^1.7.2",
     "uniqid": "^1.0.0",
-    "yargs": "^3.29.0"
+    "winston": "^2.1.1",
+    "yargs": "^3.32.0"
   },
   "devDependencies": {
-    "babel-core": "^5.8.29",
-    "babel-eslint": "^4.1.6",
-    "babel-loader": "^5.3.2",
-    "babel-runtime": "^5.8.29",
-    "browser-sync": "^2.9.6",
-    "css-loader": "^0.23.0",
+    "autoprefixer": "^6.3.2",
+    "babel-cli": "^6.5.1",
+    "babel-core": "^6.5.1",
+    "babel-eslint": "^5.0.0-beta9",
+    "babel-loader": "^6.2.2",
+    "babel-plugin-transform-runtime": "^6.5.0",
+    "babel-preset-es2015": "^6.5.0",
+    "babel-preset-react": "^6.5.0",
+    "babel-preset-react-hmre": "^1.1.0",
+    "babel-preset-stage-0": "^6.5.0",
+    "babel-register": "^6.5.1",
+    "babel-runtime": "^6.3.19",
+    "copy-webpack-plugin": "^1.1.1",
+    "css-loader": "^0.23.1",
     "eslint": "^1.10.3",
-    "eslint-config-airbnb": "^2.0.0",
-    "eslint-loader": "^1.0.0",
-    "eslint-plugin-react": "^3.11.2",
-    "extract-text-webpack-plugin": "^0.9.0",
+    "eslint-config-airbnb": "^5.0.0",
+    "eslint-loader": "^1.2.1",
+    "eslint-plugin-babel": "^3.1.0",
+    "eslint-plugin-react": "^3.16.1",
+    "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.8.5",
-    "gulp": "^3.9.0",
+    "gulp": "^3.9.1",
+    "gulp-conventional-changelog": "^1.0.1",
     "gulp-iconify": "^0.4.0",
     "gulp-load-plugins": "^1.0.0",
-    "html-webpack-plugin": "^1.7.0",
+    "html-webpack-plugin": "^2.8.1",
     "jasmine-core": "^2.3.4",
     "json-loader": "^0.5.4",
     "karma": "^0.13.14",
