@@ -1,8 +1,13 @@
 import React from 'react';
+import { connect } from 'react-redux';
 
 import Page from 'components/pages/Page';
 
-class Hello extends Page {
+const mapStateToProps = (state) => ({
+  routing: state.routing, // Available with `react-router-redux`
+  tools: state.tools,
+});
+export class Hello extends Page {
 
   getDataPage() {
     return 'hello-world';
@@ -13,9 +18,7 @@ class Hello extends Page {
       <div className="page-content">
         <div className="content-block-title">Hello World</div>
         <div className="content-block">
-          {/* If no data-panel attribute, Left panel will be opened by default */}
           <p><a href="#" className="open-panel ajax">Open Left Panel</a></p>
-          {/* Click on link with "open-panel" and data-panel="right" attribute will open Right panel */}
           <p><a href="#" data-panel="right" className="open-panel ajax">Open Right Panel</a></p>
         </div>
       </div>
@@ -23,4 +26,4 @@ class Hello extends Page {
   }
 }
 
-export default Hello;
+export default connect(mapStateToProps)(Hello);
