@@ -1,14 +1,8 @@
 import React from 'react';
-import { connect } from 'react-redux';
 
 import Page from 'components/pages/Page';
 
-@connect(
-  state => ({
-    router: state.router,
-  })
-)
-class Home extends Page {
+export default class Home extends Page {
 
   getDataPage() {
     return 'home';
@@ -21,12 +15,13 @@ class Home extends Page {
         <div className="content-block">
           {/* If no data-panel attribute, Left panel will be opened by default */}
           <p><a href="#" className="open-panel ajax">Open Left Panel</a></p>
-          {/* Click on link with "open-panel" and data-panel="right" attribute will open Right panel */}
+          {/*
+            Click on link with "open-panel" and
+            data-panel="right" attribute will open Right panel
+          */}
           <p><a href="#" data-panel="right" className="open-panel ajax">Open Right Panel</a></p>
         </div>
       </div>
     );
   }
 }
-
-export default Home;
