@@ -4,8 +4,10 @@ import LogMonitor from 'redux-devtools-log-monitor';
 import DockMonitor from 'redux-devtools-dock-monitor';
 
 export default createDevTools(
-  <DockMonitor toggleVisibilityKey="H"
-               changePositionKey="Q">
+  <DockMonitor
+    toggleVisibilityKey="H"
+    changePositionKey="Q"
+  >
     <LogMonitor />
   </DockMonitor>
 );
