@@ -1,24 +1,25 @@
 import React, { Component, PropTypes } from 'react';
 import { Provider } from 'react-redux';
-import { ReduxRouter } from 'redux-router';
+import { Router } from 'react-router';
 
 import routes from 'routes';
 
 const propTypes = {
+  history: PropTypes.object.isRequired,
   store: PropTypes.object.isRequired,
 };
 
 class RootProd extends Component {
 
   render() {
-    const { store } = this.props;
+    const { history, store } = this.props;
 
     return (
       <Provider store={store}>
         <div className="app-container">
-          <ReduxRouter>
+          <Router history={history}>
             { routes }
-          </ReduxRouter>
+          </Router>
         </div>
       </Provider>
     );
