@@ -1,7 +1,8 @@
 import React from 'react';
 import ReactDom from 'react-dom';
+import { hashHistory } from 'react-router';
 
-import configureStore from 'store';
+import configureStore from './redux/store';
 
 import Root from 'containers/Root';
 
@@ -9,21 +10,18 @@ import Root from 'containers/Root';
 import 'framework7';
 
 // *** STYLES *** //
-// Path to Framework7 MATERIAL CSS theme styles
-import 'framework7.material.min.css';
-// Path to Framework7 MATERIAL related color styles
-import 'framework7.material.colors.min.css';
 // Path to svg logos icons
-import 'assets/vendors/icons.svg.scss';
-import 'font-awesome.scss';
-import 'styles/scss/index.scss';
+import 'assets/vendors/icons.svg.css';
+import 'styles/app.css';
 
-const store = configureStore();
+const store = configureStore({}, hashHistory);
+const root = (<Root history={ hashHistory } store={ store } />);
 
 window.f7App = new Framework7({
-  swipePanel: 'left', // http://www.idangero.us/framework7/docs/side-panels.html#open-panels-with-swipe
+  // http://www.idangero.us/framework7/docs/side-panels.html#open-panels-with-swipe
+  swipePanel: 'left',
   ajaxLinks: 'a.ajax',
   material: true,
 });
 
-ReactDom.render(<Root store={ store } />, document.getElementById('app'));
+ReactDom.render(root, document.getElementById('app'));
