@@ -5,24 +5,25 @@
   // - https://github.com/rackt/react-router/blob/master/docs/guides/basics/Histories.md
 
 import React from 'react';
-import { Route, IndexRoute } from 'react-router';
+import { Route, IndexRoute, Redirect } from 'react-router';
 
-import App from 'containers/App';
+import AppLayout from 'layouts/AppLayout';
 import Home from 'components/pages/Home';
 import About from 'components/pages/About';
 import Hello from 'components/pages/Hello';
 import Counter from 'components/pages/Counter';
 import Tools from 'components/pages/Tools';
+import NotFound from 'components/pages/NotFound';
 
 export default (
-  <Route path="/" component={App}>
+  <Route path="/" component={AppLayout}>
     <IndexRoute component={Home} />
     <Route path="home" component={Home} />
     <Route path="hello" component={Hello} />
     <Route path="about" component={About} />
     <Route path="counter" component={Counter} />
     <Route path="tools" component={Tools} />
-    { /* https://github.com/rackt/react-router/blob/master/UPGRADE_GUIDE.md#notfound-route */ }
-    <Route path="*" component={Home}/>
+    <Route path="404" component={NotFound}/>
+    <Redirect from="*" to="404" />
   </Route>
 );
